+++ Editor/Migration/V2/AtlasTextureV2.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using net.rs64.TexTransTool.TextureAtlas;
 using UnityEditor;
 using UnityEngine;
@@ -14,6 +15,7 @@ namespace net.rs64.TexTransTool.Migration.V2
             if (atlasTexture is ITexTransToolTag TTTag && TTTag.SaveDataVersion > 3) { Debug.Log(atlasTexture.name + " AtlasTexture : マイグレーション不可能なバージョンです。"); return; }
 
             atlasTexture.AtlasSetting.IslandPadding = atlasTexture.AtlasSetting.Padding / atlasTexture.AtlasSetting.AtlasTextureSize;
+            atlasTexture.AtlasSetting.TextureFineTuning = atlasTexture.AtlasSetting.TextureFineTuningDataList.Select(i => i.GetFineTuning()).ToList();
 
             EditorUtility.SetDirty(atlasTexture);
             MigrationUtility.SetSaveDataVersion(atlasTexture, 3);